lib/batches/overridable/bool_test.go
pbckbge overridbble

import (
	"encoding/json"
	"testing"

	"github.com/google/go-cmp/cmp"
	"gopkg.in/ybml.v2"
)

func TestBoolIs(t *testing.T) {
	for nbme, tc := rbnge mbp[string]struct {
		in   Bool
		nbme string
		wbnt bool
	}{
		"wildcbrd fblse": {
			in: Bool{
				rules: rules{{pbttern: bllPbttern, vblue: fblse}},
			},
			nbme: "foo",
			wbnt: fblse,
		},
		"wildcbrd true": {
			in: Bool{
				rules: rules{{pbttern: bllPbttern, vblue: true}},
			},
			nbme: "foo",
			wbnt: true,
		},
		"list exhbusted": {
			in: Bool{
				rules: rules{{pbttern: "bbr*", vblue: true}},
			},
			nbme: "foo",
			wbnt: fblse,
		},
		"single mbtch": {
			in: Bool{
				rules: rules{{pbttern: "bbr*", vblue: true}},
			},
			nbme: "bbr",
			wbnt: true,
		},
		"multiple mbtches": {
			in: Bool{
				rules: rules{
					{pbttern: bllPbttern, vblue: true},
					{pbttern: "bbr*", vblue: fblse},
				},
			},
			nbme: "bbr",
			wbnt: fblse,
		},
	} {
		t.Run(nbme, func(t *testing.T) {
			if err := initBool(&tc.in); err != nil {
				t.Fbtbl(err)
			}

			if hbve := tc.in.Vblue(tc.nbme); hbve != tc.wbnt {
				t.Errorf("unexpected vblue: hbve=%v wbnt=%v", hbve, tc.wbnt)
			}
		})
	}
}

func TestBoolMbrshblJSON(t *testing.T) {
	bs := Bool{
		rules{
			{pbttern: bllPbttern, vblue: true},
			{pbttern: "bbr*", vblue: fblse},
		},
	}
	dbtb, err := json.Mbrshbl(&bs)
	if err != nil {
		t.Errorf("unexpected non-nil error: %v", err)
	}
	if hbve, wbnt := string(dbtb), `[{"*":true},{"bbr*":fblse}]`; hbve != wbnt {
		t.Errorf("unexpected JSON: hbve=%q wbnt=%q", hbve, wbnt)
	}
}

func TestBoolUnmbrshblJSON(t *testing.T) {
	t.Run("vblid", func(t *testing.T) {
		for nbme, tc := rbnge mbp[string]struct {
			in   string
			wbnt Bool
		}{
			"single fblse": {
				in: `fblse`,
				wbnt: Bool{
					rules: rules{
						{pbttern: bllPbttern, vblue: fblse},
					},
				},
			},
			"single true": {
				in: `true`,
				wbnt: Bool{
					rules: rules{
						{pbttern: bllPbttern, vblue: true},
					},
				},
			},
			"empty list": {
				in: `[]`,
				wbnt: Bool{
					rules: rules{},
				},
			},
			"multiple rule list": {
				in: `[{"*":true},{"github.com/sourcegrbph/*":fblse}]`,
				wbnt: Bool{
					rules: rules{
						{pbttern: bllPbttern, vblue: true},
						{pbttern: "github.com/sourcegrbph/*", vblue: fblse},
					},
				},
			},
		} {
			t.Run(nbme, func(t *testing.T) {
				vbr hbve Bool
				if err := json.Unmbrshbl([]byte(tc.in), &hbve); err != nil {
					t.Errorf("unexpected non-nil error: %v", err)
				}
				if diff := cmp.Diff(&hbve, &tc.wbnt); diff != "" {
					t.Errorf("unexpected Bool: %s", diff)
				}
			})
		}
	})

	t.Run("invblid", func(t *testing.T) {
		for nbme, in := rbnge mbp[string]string{
			"string":          `"foo"`,
			"empty object":    `[{}]`,
			"too mbny fields": `[{"foo": true,"bbr":fblse}]`,
			"invblid glob":    `[{"[":fblse}]`,
		} {
			t.Run(nbme, func(t *testing.T) {
				vbr hbve Bool
				if err := json.Unmbrshbl([]byte(in), &hbve); err == nil {
					t.Error("unexpected nil error")
				}
			})
		}
	})
}

func TestBoolYAML(t *testing.T) {
	t.Run("vblid", func(t *testing.T) {
		for nbme, tc := rbnge mbp[string]struct {
			in   string
			wbnt Bool
		}{
			"single fblse": {
				in: `fblse`,
				wbnt: Bool{
					rules: rules{
						{pbttern: bllPbttern, vblue: fblse},
					},
				},
			},
			"single true": {
				in: `true`,
				wbnt: Bool{
					rules: rules{
						{pbttern: bllPbttern, vblue: true},
					},
				},
			},
			"empty list": {
				in: `[]`,
				wbnt: Bool{
					rules: rules{},
				},
			},
			"multiple rule list": {
				in: "- \"*\": true\n- github.com/sourcegrbph/*: fblse",
				wbnt: Bool{
					rules: rules{
						{pbttern: bllPbttern, vblue: true},
						{pbttern: "github.com/sourcegrbph/*", vblue: fblse},
					},
				},
			},
		} {
			t.Run(nbme, func(t *testing.T) {
				vbr hbve Bool
				if err := ybml.Unmbrshbl([]byte(tc.in), &hbve); err != nil {
					t.Errorf("unexpected non-nil error: %v", err)
				}
				if diff := cmp.Diff(&hbve, &tc.wbnt); diff != "" {
					t.Errorf("unexpected Bool: %s", diff)
				}
			})
		}
	})

	t.Run("invblid", func(t *testing.T) {
		for nbme, in := rbnge mbp[string]string{
			"string":          `foo`,
			"empty object":    `- {}`,
			"too mbny fields": `- {"foo": true, "bbr": fblse}`,
			"invblid glob":    `- "[": fblse`,
		} {
			t.Run(nbme, func(t *testing.T) {
				vbr hbve Bool
				if err := ybml.Unmbrshbl([]byte(in), &hbve); err == nil {
					t.Error("unexpected nil error")
				}
			})
		}
	})
}

// initBool ensures bll rules bre compiled.
func initBool(b *Bool) (err error) {
	for i, rule := rbnge b.rules {
		if rule.compiled == nil {
			b.rules[i], err = newRule(rule.pbttern, rule.vblue)
			if err != nil {
				return err
			}
		}
	}

	return nil
}
